better webp rewrite test, webp-only bulk mode, and web bulk notice when enabling conversion

+++ unique.php
@@ -405,7 +405,8 @@ function ewww_image_optimizer( $file, $gallery_type = 4, $converted = false, $ne
 			if (
 				1 === (int) $gallery_type &&
 				$fullsize &&
-				( ewww_image_optimizer_get_option( 'ewww_image_optimizer_jpg_to_png' ) || ! empty( $_REQUEST['ewww_convert'] ) )
+				( ewww_image_optimizer_get_option( 'ewww_image_optimizer_jpg_to_png' ) || ! empty( $_REQUEST['ewww_convert'] ) ) &&
+				empty( $_REQUEST['ewww_webp_only'] )
 			) {
 				// Generate the filename for a PNG.
 				if ( $converted ) { // If this is a resize version...
@@ -428,20 +429,21 @@ function ewww_image_optimizer( $file, $gallery_type = 4, $converted = false, $ne
 				}
 			}
 			$ewww_image->level = (int) ewww_image_optimizer_get_option( 'ewww_image_optimizer_jpg_level' );
-			if ( ewww_image_optimizer_get_option( 'ewww_image_optimizer_jpg_level' ) > 0 ) {
+			if ( ewww_image_optimizer_get_option( 'ewww_image_optimizer_jpg_level' ) > 0 && empty( $_REQUEST['ewww_webp_only'] ) ) {
 				list( $file, $converted, $result, $new_size, $backup_hash ) = ewww_image_optimizer_cloud_optimizer( $file, $type, $convert, $pngfile, 'image/png', $skip_lossy );
 				if ( $converted ) {
 					if ( ewww_image_optimizer_get_option( 'ewww_image_optimizer_delete_originals' ) ) {
 						// Delete the original JPG.
 						unlink( $original );
 					}
-					$converted = $filenum;
-					ewww_image_optimizer_webp_create( $file, $new_size, 'image/png', null );
+					$converted   = $filenum;
+					$webp_result = ewww_image_optimizer_webp_create( $file, $new_size, 'image/png', null, $orig_size !== $new_size );
 				} else {
-					ewww_image_optimizer_webp_create( $file, $new_size, $type, null );
+					$webp_result = ewww_image_optimizer_webp_create( $file, $new_size, $type, null, $orig_size !== $new_size );
 				}
 			} else {
-				ewww_image_optimizer_webp_create( $file, $orig_size, $type, null );
+				ewwwio_debug_message( 'calling webp, but neither convert or optimize' );
+				$webp_result = ewww_image_optimizer_webp_create( $file, $orig_size, $type, null );
 			}
 			break;
 		case 'image/png':
@@ -451,7 +453,8 @@ function ewww_image_optimizer( $file, $gallery_type = 4, $converted = false, $ne
 				1 === (int) $gallery_type &&
 				$fullsize &&
 				( ewww_image_optimizer_get_option( 'ewww_image_optimizer_png_to_jpg' ) || ! empty( $_REQUEST['ewww_convert'] ) ) &&
-				! $skip_lossy
+				! $skip_lossy &&
+				empty( $_REQUEST['ewww_webp_only'] )
 			) {
 				ewwwio_debug_message( 'PNG to JPG conversion turned on' );
 				// If the user set a fill background for transparency.
@@ -490,20 +493,21 @@ function ewww_image_optimizer( $file, $gallery_type = 4, $converted = false, $ne
 				}
 			}
 			$ewww_image->level = (int) ewww_image_optimizer_get_option( 'ewww_image_optimizer_png_level' );
-			if ( ewww_image_optimizer_get_option( 'ewww_image_optimizer_png_level' ) > 0 ) {
+			if ( ewww_image_optimizer_get_option( 'ewww_image_optimizer_png_level' ) > 0 && empty( $_REQUEST['ewww_webp_only'] ) ) {
 				list( $file, $converted, $result, $new_size, $backup_hash ) = ewww_image_optimizer_cloud_optimizer( $file, $type, $convert, $jpgfile, 'image/jpeg', $skip_lossy, $cloud_background, $quality );
 				if ( $converted ) {
 					if ( ewww_image_optimizer_get_option( 'ewww_image_optimizer_delete_originals' ) ) {
 						// Delete the original JPG.
 						unlink( $original );
 					}
-					$converted = $filenum;
-					ewww_image_optimizer_webp_create( $file, $new_size, 'image/jpeg', null );
+					$converted   = $filenum;
+					$webp_result = ewww_image_optimizer_webp_create( $file, $new_size, 'image/jpeg', null, $orig_size !== $new_size );
 				} else {
-					ewww_image_optimizer_webp_create( $file, $new_size, $type, null );
+					$webp_result = ewww_image_optimizer_webp_create( $file, $new_size, $type, null, $orig_size !== $new_size );
 				}
 			} else {
-				ewww_image_optimizer_webp_create( $file, $orig_size, $type, null );
+				ewwwio_debug_message( 'calling webp, but neither convert or optimize' );
+				$webp_result = ewww_image_optimizer_webp_create( $file, $orig_size, $type, null );
 			}
 			break;
 		case 'image/gif':
@@ -543,8 +547,8 @@ function ewww_image_optimizer( $file, $gallery_type = 4, $converted = false, $ne
 						// Delete the original JPG.
 						unlink( $original );
 					}
-					$converted = $filenum;
-					ewww_image_optimizer_webp_create( $file, $new_size, 'image/png', null );
+					$converted   = $filenum;
+					$webp_result = ewww_image_optimizer_webp_create( $file, $new_size, 'image/png', null );
 				}
 			}
 			break;
@@ -584,10 +588,16 @@ function ewww_image_optimizer( $file, $gallery_type = 4, $converted = false, $ne
 			copy( $file, $s3_uploads_image );
 		}
 		$file = ewww_image_optimizer_s3_uploads_image_cleanup( $file );
+		if ( ! empty( $webp_result ) ) {
+			$results_msg .= '<br>' . $webp_result;
+		}
 		ewwwio_memory( __FUNCTION__ );
 		return array( $file, $results_msg, $converted, $original );
 	} else {
 		ewww_image_optimizer_s3_uploads_image_cleanup( $file );
+	}
+	if ( ! empty( $webp_result ) && ! empty( $_REQUEST['ewww_webp_only'] ) ) {
+		$result = $webp_result;
 	}
 	ewwwio_memory( __FUNCTION__ );
 	// otherwise, send back the filename, the results (some sort of error message), the $converted flag, and the name of the original image.
@@ -595,13 +605,14 @@ function ewww_image_optimizer( $file, $gallery_type = 4, $converted = false, $ne
 }
 
 /**
- * Creates webp images alongside JPG and PNG files.
+ * Creates WebP images alongside JPG and PNG files.
  *
  * @param string $file The name of the JPG/PNG file.
  * @param int    $orig_size The filesize of the JPG/PNG file.
  * @param string $type The mime-type of the incoming file.
  * @param string $tool The path to the cwebp binary, if installed.
  * @param bool   $recreate True to keep the .webp image even if it is larger than the JPG/PNG.
+ * @return string Results of the WebP operation for display.
  */
 function ewww_image_optimizer_webp_create( $file, $orig_size, $type, $tool, $recreate = false ) {
 	ewwwio_debug_message( '<b>' . __FUNCTION__ . '()</b>' );
@@ -609,12 +620,12 @@ function ewww_image_optimizer_webp_create( $file, $orig_size, $type, $tool, $rec
 	$bypass_webp = apply_filters( 'ewww_image_optimizer_bypass_webp', false, $file );
 	if ( true === $bypass_webp ) {
 		ewwwio_debug_message( "webp generation bypassed: $file" );
-		return;
+		return '';
 	} elseif ( ! ewww_image_optimizer_get_option( 'ewww_image_optimizer_webp' ) ) {
-		return;
+		return '';
 	} elseif ( is_file( $webpfile ) && empty( $_REQUEST['ewww_force'] ) && ! $recreate ) {
 		ewwwio_debug_message( 'webp file exists, not forcing or recreating' );
-		return;
+		return esc_html__( 'WebP image already exists.', 'ewww-image-optimizer-cloud' );
 	}
 	ewww_image_optimizer_cloud_optimizer( $file, $type, false, $webpfile, 'image/webp' );
 	$webp_size = ewww_image_optimizer_filesize( $webpfile );
@@ -622,13 +633,18 @@ function ewww_image_optimizer_webp_create( $file, $orig_size, $type, $tool, $rec
 	if ( is_file( $webpfile ) && $orig_size < $webp_size && ! ewww_image_optimizer_get_option( 'ewww_image_optimizer_webp_force' ) ) {
 		ewwwio_debug_message( 'webp file was too big, deleting' );
 		unlink( $webpfile );
+		return esc_html__( 'WebP image was larger than original.', 'ewww-image-optimizer-cloud' );
 	} elseif ( is_file( $webpfile ) ) {
 		// Set correct file permissions.
 		$stat  = stat( dirname( $webpfile ) );
 		$perms = $stat['mode'] & 0000666; // same permissions as parent folder, strip off the executable bits.
 		chmod( $webpfile, $perms );
+		if ( $orig_size < $webp_size && ewww_image_optimizer_get_option( 'ewww_image_optimizer_webp_force' ) ) {
+			return esc_html__( 'WebP image larger than original, saved anyway with Force WebP option.', 'ewww-image-optimizer-cloud' );
+		}
+		return 'WebP: ' . ewww_image_optimizer_image_results( $orig_size, $webp_size );
 	}
-	ewwwio_memory( __FUNCTION__ );
+	return esc_html__( 'Image could not be converted to WebP.', 'ewww-image-optimizer-cloud' );
 }
 
 /**
